@@ -63,65 +63,66 @@
     "yamlify": "^0.1.3"
   },
   "devDependencies": {
-    "autoprefixer": "^6.0.3",
+    "autoprefixer": "^6.7.2",
     "babel-eslint": "^7.1.1",
-    "babel-loader": "^6.2.7",
+    "babel-loader": "^6.2.10",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-export-extensions": "^6.4.0",
-    "babel-plugin-transform-react-constant-elements": "^6.4.0",
-    "babel-plugin-transform-react-inline-elements": "^6.8.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.2.11",
-    "babel-plugin-transform-runtime": "^6.15.0",
-    "babel-preset-latest": "^6.16.0",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-stage-2": "^6.18.0",
-    "babel-register": "^6.7.2",
-    "chai": "^3.4.1",
+    "babel-plugin-transform-export-extensions": "^6.22.0",
+    "babel-plugin-transform-react-constant-elements": "^6.22.0",
+    "babel-plugin-transform-react-inline-elements": "^6.22.0",
+    "babel-plugin-transform-react-remove-prop-types": "^0.3.0",
+    "babel-plugin-transform-runtime": "^6.22.0",
+    "babel-preset-latest": "^6.22.0",
+    "babel-preset-react": "^6.22.0",
+    "babel-preset-stage-2": "^6.22.0",
+    "babel-register": "^6.22.0",
+    "chai": "^3.5.0",
     "cheerio": "^0.22.0",
-    "cross-env": "^3.1.3",
-    "css-loader": "^0.26.0",
-    "cssnano": "^3.3.1",
-    "del": "^2.0.2",
-    "eslint": "^3.13.0",
+    "cross-env": "^3.1.4",
+    "css-loader": "^0.26.1",
+    "cssnano": "^3.10.0",
+    "del": "^2.2.2",
+    "eslint": "^3.15.0",
     "eslint-config-airbnb": "^14.0.0",
     "eslint-config-airbnb-base": "^11.0.1",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^3.0.2",
     "eslint-plugin-react": "^6.9.0",
-    "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^2.0.0-beta.4",
-    "fetch-mock": "^5.0.3",
-    "gulp": "^3.9.0",
-    "gulp-babel": "^6.1.1",
+    "express": "^4.14.1",
+    "extract-text-webpack-plugin": "^2.0.0-rc.3",
+    "fetch-mock": "^5.9.3",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
     "gulp-htmlmin": "^3.0.0",
-    "gulp-if": "^2.0.0",
-    "gulp-image": "^2.7.2",
+    "gulp-if": "^2.0.2",
+    "gulp-image": "^2.7.3",
     "gulp-insert": "^0.5.0",
-    "gulp-newer": "^1.1.0",
+    "gulp-newer": "^1.3.0",
     "gulp-plumber": "^1.1.0",
-    "gulp-util": "^3.0.7",
-    "gulp-watch": "^4.3.9",
+    "gulp-util": "^3.0.8",
+    "gulp-watch": "^4.3.11",
     "gulp-yaml": "^1.0.1",
-    "html-webpack-plugin": "^2.24.1",
+    "html-webpack-plugin": "^2.28.0",
     "json-loader": "^0.5.4",
-    "loud-rejection": "^1.3.0",
-    "mocha": "^3.0.0",
-    "npm-run-path": "^2.0.1",
-    "postcss-cssnext": "^2.8.0",
-    "postcss-import": "^9.0.0",
-    "postcss-loader": "^1.1.1",
-    "proxyquire": "^1.7.4",
-    "run-sequence": "^1.2.0",
-    "sinon": "^1.17.5",
+    "lodash-webpack-plugin": "^0.10.6",
+    "loud-rejection": "^1.6.0",
+    "mocha": "^3.2.0",
+    "npm-run-path": "^2.0.2",
+    "postcss-cssnext": "^2.9.0",
+    "postcss-import": "^9.1.0",
+    "postcss-loader": "^1.2.2",
+    "proxyquire": "^1.7.11",
+    "run-sequence": "^1.2.2",
+    "sinon": "^1.17.7",
     "sinon-chai": "^2.8.0",
     "style-loader": "^0.13.1",
-    "through2": "^2.0.1",
-    "u-wave-source-soundcloud": "^1.0.0",
-    "u-wave-source-youtube": "^1.0.0",
+    "through2": "^2.0.3",
+    "u-wave-source-soundcloud": "^1.0.1",
+    "u-wave-source-youtube": "^1.0.4",
     "u-wave-web-emojione": "^1.0.0",
-    "webpack": "^2.1.0-beta.27",
-    "webpack-dev-middleware": "^1.8.4",
-    "webpack-hot-middleware": "^2.13.2",
+    "webpack": "^2.2.1",
+    "webpack-dev-middleware": "^1.10.0",
+    "webpack-hot-middleware": "^2.16.1",
     "yaml-loader": "^0.4.0"
   },
   "scripts": {
