@@ -30,15 +30,6 @@ export default class ChatMessages extends React.Component {
     }
   }
 
-  makeDeleteHandler(msg) {
-    if (this.props.canDeleteMessages) {
-      return () => {
-        this.props.onDeleteMessage(msg._id);
-      };
-    }
-    return null;
-  }
-
   scrollToBottom() {
     const el = this.container;
     el.scrollTop = el.scrollHeight;
@@ -69,13 +60,11 @@ export default class ChatMessages extends React.Component {
     );
   }
 
-  renderMessage(msg, index) {
-    const alternate = index % 2 === 0;
+  renderMessage(msg) {
     if (msg.type === 'log') {
       return (
         <LogMessage
           key={msg._id}
-          alternate={alternate}
           {...msg}
         />
       );
@@ -84,9 +73,8 @@ export default class ChatMessages extends React.Component {
     return (
       <Message
         key={msg._id}
-        alternate={alternate}
         compileOptions={this.props.compileOptions}
-        onDelete={this.makeDeleteHandler(msg)}
+        onDelete={this.props.canDeleteMessages ? this.props.onDeleteMessage : null}
         {...msg}
       />
     );
