@@ -1,15 +1,8 @@
-import cx from 'classnames';
 import * as React from 'react';
 import pure from 'recompose/pure';
 
-const LogMessage = ({ alternate, text }) => (
-  <div
-    className={cx(
-      'ChatMessage',
-      'ChatMessage--log',
-      alternate && 'ChatMessage--alternate'
-    )}
-  >
+const LogMessage = ({ text }) => (
+  <div className="ChatMessage ChatMessage--log">
     <div className="ChatMessage-content">
       <span className="ChatMessage-text">{text}</span>
     </div>
@@ -17,7 +10,6 @@ const LogMessage = ({ alternate, text }) => (
 );
 
 LogMessage.propTypes = {
-  alternate: React.PropTypes.bool,
   text: React.PropTypes.string.isRequired
 };
 
